+++ libs/stack/stack-ui/src/components/Alerts/components/pagination/AlertsPaginationBullet.tsx
@@ -1,12 +1,38 @@
-import useThemeContext from 'libs/stack/stack-ui/src/providers/Theme/hooks'
+import { useAlertsController } from 'libs/stack/stack-ui/src/providers/Alerts'
+import { useFocusManager, useKeyboard, usePress } from 'react-aria'
 import slugify from 'slugify'
+import Button from '../../../Button'
 import type { TAlertsPaginationBulletProps } from '../../interface'
 
 const AlertsPaginationBullet = (props: TAlertsPaginationBulletProps) => {
-  const { themeName, tokens, controller, alerts, activeIndex, index } = props
+  const { themeName, tokens, alerts, activeIndex, index } = props
 
-  const paginationBulletTheme = useThemeContext(`${themeName}.bullet`, tokens)
-  const paginationActiveBulletTheme = useThemeContext(`${themeName}.activeBullet`, tokens)
+  const { controller } = useAlertsController()
+
+  const focusManager = useFocusManager()
+  const prevIndex = index === 0 ? alerts.length - 1 : index - 1
+  const nextIndex = index === alerts.length - 1 ? 0 : index + 1
+
+  const { keyboardProps } = useKeyboard({
+    onKeyUp: (e) => {
+      if (e.key === 'ArrowLeft') {
+        e.preventDefault()
+        focusManager?.focusPrevious({ wrap: true })
+        controller?.slideTo(prevIndex)
+      }
+
+      if (e.key === 'ArrowRight') {
+        e.preventDefault()
+        focusManager?.focusNext({ wrap: true })
+        controller?.slideTo(nextIndex)
+      }
+    },
+  })
+
+  const { pressProps } = usePress({
+    onPress: () => controller?.slideTo(index),
+  })
+
   const alert = alerts[index]
   const isActive = index === activeIndex
   const { title, id } = alert
@@ -14,22 +40,20 @@ const AlertsPaginationBullet = (props: TAlertsPaginationBulletProps) => {
   const hasTitle = title && title.length > 0
 
   return (
-    <span
-      tabIndex={0}
-      role="button"
-      onKeyDown={(event) => {
-        if (event.key === 'Enter' || event.key === ' ') {
-          event.preventDefault()
-          controller?.slideTo(index)
-        }
+    <Button
+      as="button"
+      {...{
+        tabIndex: 0,
+        role: 'button',
+        'aria-current': isActive ? 'true' : 'false',
+        'aria-disabled': isActive,
+        'aria-label': !hasTitle ? `${index + 1} / ${alerts.length}` : undefined,
+        'aria-labelledby': hasTitle ? slugify(`${id}-${title}`) : undefined,
       }}
-      onClick={() => controller?.slideTo(index)}
-      key={JSON.stringify(alert)}
-      className={`${paginationBulletTheme} ${isActive ? paginationActiveBulletTheme : ''}`}
-      aria-current={isActive ? 'true' : 'false'}
-      aria-label={!hasTitle ? `${index + 1} / ${alerts.length}` : undefined}
-      aria-disabled={isActive}
-      aria-labelledby={hasTitle ? slugify(`${id}-${title}`) : undefined}
+      {...keyboardProps}
+      {...pressProps}
+      themeName={`${themeName}.bullet`}
+      tokens={{ ...tokens, active: isActive }}
     />
   )
 }